+++ packages/desktop/.gitignore
@@ -1,0 +1,15 @@
+.DS_Store
+.env.development.local
+.env.local
+.env.production.local
+.env.test.local
+.history
+.jest
+.vscode
+build
+coverage
+dist
+node_modules
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
