+++ packages/desktop/tslint.json
@@ -1,0 +1,3 @@
+{
+  "extends": ["../../tslint.json"]
+}
